binary data parsing 4.1.1, 4.1.2 & 4.1.3 OK

+++ sound-data-converter/src/main/java/com/gaia3d/utils/io/LittleEndianDataInputStream.java
@@ -187,7 +187,7 @@ public class LittleEndianDataInputStream extends FilterInputStream implements Da
         ByteBuffer byteBuffer = ByteBuffer
                 .wrap(bytes)
                 .order(ByteOrder.LITTLE_ENDIAN);
-        float result = byteBuffer.getFloat();
+        double result = byteBuffer.getDouble();
         byteBuffer.clear();
         return result;
     }
